Fixing a bunch.

+++ output/latex/latex.c
@@ -131,7 +131,7 @@ static void WriteFixElement(FILE *file,
                             char const *name, char const *value,
                             unsigned int indentation)
 {
-  fprintf(file,"%*c%s{%s}%%\n",indentation+1,'\\',name,value);
+  fprintf(file,"%*c%s{%s}%%\n",(int)(indentation+1),'\\',name,value);
 }
 
 void WriteUserInputElement(FILE *file, char const *name, char const *value)
@@ -145,7 +145,7 @@ static void WriteUserInputSubElement(FILE *file,
                                      char const *name,
                                      unsigned value_length, char const *value)
 {
-  fprintf(file," \\%s{%.*s}%%\n",name,value_length,value);
+  fprintf(file," \\%s{%.*s}%%\n",name,(int)value_length,value);
 }
 
 static void WriteGeneratedElement(FILE *file,
